Extract adaptive staging conversion and separate drain and pressure code
